% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{epi}
\alias{epi}
\title{An interface to eight compartmental epidemic models.}
\usage{
epi(
  population,
  susceptible,
  exposed = 0,
  infected,
  recovered,
  dead = 0,
  beta,
  gamma,
  sigma = 0,
  delta = 0,
  xi = 0,
  muBirth = 0,
  muDeath = 0,
  vitalDynamics = FALSE,
  trueMassAction = FALSE,
  timesteps = 25,
  timeIncrement = 1
)
}
\arguments{
\item{population}{The initial number of persons in the simulation; if
\code{vitalDynamics = TRUE} and \code{muBirth} and \code{muDeath} are equal, then it is constant.}

\item{susceptible}{The initial number of people who are susceptible to the
disease (they have no natural immunity to it).}

\item{exposed}{The initial number of people who have been exposed to the
infection and may become infectious themselves.}

\item{infected}{The initial number of people who are infectious and may
infect others.}

\item{recovered}{The initial number of people whom are no longer infectious
and not a risk to others.}

\item{dead}{The initial number of fatalities (people killed by the disease).}

\item{beta}{The rate of becoming infectious (if \code{sigma = 0}, the default), or
the rate of susceptible people being exposed to the infectious.}

\item{gamma}{The rate of becoming recovered (no longer infectious, if \code{sigma = 0}, the default), or the rate of becoming infectious after exposure.}

\item{sigma}{The rate of becoming recovered (no longer infectious).}

\item{delta}{The rate of fatality.}

\item{xi}{The rate of loss of learned immunity.}

\item{muBirth}{The rate of birth.}

\item{muDeath}{The rate of death not due to the disease under consideration.}

\item{vitalDynamics}{Whether vital dynamics is enabled (if \code{FALSE}, the
default, implies muBirth and muDeath of zero).}

\item{trueMassAction}{Whether the force of infection is scaled by the
population or not.}

\item{timesteps}{The number of output rows in the simulation.}

\item{timeIncrement}{The number of increments of "time" that pass between
each output row.}
}
\value{
A dataframe with a column for each enabled compartment and a column
for time, with \code{timesteps} rows. Each row contains the values of the
differential equations for each point in time of the model (per function
arguments).
}
\description{
ehpi is an application programming interface (for the Shiny application that
motivated its creation) or a convenient function \code{\link[=epi]{epi()}} for
interactive use in epidemic modelling. The arguments provided to the function
determine compartments used in the model that is run.
}
\details{
The convention for pronouncing \code{\link[=epi]{epi()}} is "epi epi" as if you were
saying "epidemic" twice. If you do want to refer to the package alone, and
not the singular function it packages, you can say "eh pee eye", like "API".

"eh pee eye" is a tongue-in-cheek Canadianism. You might understand it as,
"Eh, P.I.?", or just a Canadian slant on API. The latter was thought of
first.

While ehpi provides "an API" for the
\href{https://github.com/ashokkrish/episim}{Episim} Shiny application, it won't
enable the development of applications that provide any feature episim
doesn't already provide by wrapping this package. The function is packages
was extracted from that project to allow associating well-formatted
documentation, vignettes, and examples with the function because that
function is expected to be used outside of Episim, interactively, during
workshops run by Dr. Krishnamurthy.
}
\examples{
epi(population = 500,
    susceptible = 489,
    exposed = 10,
    infected = 1,
    recovered = 0,
    dead = 29,
    beta = 0.005,
    gamma = 0.0009,
    sigma = 0.000001,
    delta = 0.01,
    timesteps = 230)

}
